add focused tab tracking; clean up ktab demo
